requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==43.0.0
distlib==0.3.8
Django==3.2.25
django-cors-headers==4.4.0
django-countries==7.6.1
django-extensions==3.2.3
django-filter==23.5
django-phonenumber-field==8.0.0
django-sslserver==0.22
djangorestframework==3.15.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.15.4
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.25.2
identify==2.6.0
idna==3.7
importlib_metadata==8.0.0
Markdown==3.6
msgpack==1.0.8
nodeenv==1.9.1
packaging==24.1
phonenumbers==8.13.43
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.7.1
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.4
pyuploadcare==5.1.0
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
ruff==0.5.4
rules==3.4
sentry-sdk==2.10.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uuid==1.30
virtualenv==20.26.3
whitenoise==6.7.0
zipp==3.19.2
